shader: move int and illegal instructions to separate files
